requirements.txt
fastapi
uvicorn[standard]==0.30.1
python-multipart==0.0.9
slowapi==0.1.9
mistralai==1.9.6
arize-phoenix-evals
requests==2.32.3
pydub==0.25.1
supabase
pytest==8.3.2
pytest-cov==5.0.0
httpx
python-dotenv==1.0.1
psycopg[binary]==3.1.19
google-genai==1.30.0
langgraph
langchain
redis
openai
sentence-transformers
ragas
pgvector
